requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
Flask-Cors
bcrypt
pandas
scikit-learn
tensorflow
numpy
python-dotenv
Werkzeug
email-validator
marshmallow
pytest
gunicorn
Pillow
